+++ examples/react/hello-world/eslint.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-    plugins: [require.resolve('@reshadow/eslint')],
-};