+++ examples/react/hello-world/prettier.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-    tabWidth: 4,
-    singleQuote: true,
-    trailingComma: 'all',
-    bracketSpacing: false,
-    plugins: [require.resolve('@reshadow/prettier')],
-};