+++ packages/webpack/loader.js
@@ -1,6 +1,9 @@
 const path = require('path');
 const os = require('os');
 const findCacheDir = require('find-cache-dir');
+const loaderUtils = require('loader-utils');
+const mkdirp = require('mkdirp');
+const fs = require('fs');
 
 const VirtualModulesPlugin = require('webpack-virtual-modules');
 
@@ -9,63 +12,122 @@ const utils = require('@reshadow/utils');
 const virtualModules = new VirtualModulesPlugin();
 
 let cacheDirectory;
+let options;
 
-module.exports = function(source) {
-    cacheDirectory =
-        cacheDirectory || findCacheDir({name: 'reshadow'}) || os.tmpdir();
+const makeLoader = loader =>
+    function(...args) {
+        // Make the loader async
+        const callback = this.async();
 
+        loader
+            .apply(this, args)
+            .then(args => callback(null, ...args), err => callback(err));
+    };
+
+async function loader(source, inputSourceMap) {
     if (this.cacheable) this.cacheable();
 
-    const {compiler} = this._compilation;
+    options =
+        options ||
+        Object.assign(
+            {
+                getFilepath: filepath => {
+                    const hash = `${utils.getFileHash(filepath)}_${++index}`;
+                    const filename = `${hash}.css`;
 
-    /**
-     * We need to tap 'after-environment' hook by hands,
-     * because there is no 'official' register for the plugin
-     */
-    compiler.hooks.afterEnvironment.intercept({
-        name: 'VirtualModulesPlugin',
-        context: true,
-        register: tap => (tap.fn(), tap),
-    });
+                    return path.resolve(cacheDirectory, filename);
+                },
+                virtualFS: true,
+            },
+            loaderUtils.getOptions(this),
+        );
 
-    virtualModules.apply(compiler);
+    cacheDirectory =
+        cacheDirectory ||
+        options.cacheDirectory ||
+        findCacheDir({name: 'reshadow'}) ||
+        os.tmpdir();
 
-    const filepath = this.resourcePath;
+    let writeModule;
+
+    if (options.virtualFS) {
+        const {compiler} = this._compilation;
+
+        /**
+         * We need to tap 'after-environment' hook by hands,
+         * because there is no 'official' register for the plugin
+         */
+        compiler.hooks.afterEnvironment.intercept({
+            name: 'VirtualModulesPlugin',
+            context: true,
+            register: tap => (tap.fn(), tap),
+        });
+
+        virtualModules.apply(compiler);
+
+        writeModule = (filepath, content) => {
+            virtualModules.writeModule(filepath, content);
+        };
+    } else {
+        writeModule = (filepath, content) => {
+            return new Promise(resolve => {
+                mkdirp(path.dirname(filepath), () => {
+                    fs.writeFile(filepath, content, resolve);
+                });
+            });
+        };
+    }
+
+    const {resourcePath} = this;
 
     let index = 0;
+
+    const queue = [];
 
     const result = source
         .replace(
             // Regexp below should match on the CSS code from strings like that:
             //
-            // __css__(`button {
-            //   /* Some CSS rules here... */
-            //   content: "*"; /* With some quotes maybe */
-            // }`
-            // /*__css_end__*/
+            // /*__reshadow_css_start__*/
+            // __css__(
+            //   /*__inner_css_start__*/
+            //   `button {
+            //     /* Some CSS rules here... */
+            //     content: "*"; /* With some quotes maybe */
+            //   }`
+            //   /*__inner_css_end__*/
             // , "2845693891")
+            // /*__reshadow_css_end__*/
             //
-            // We're using trailing block comment /*__css_end__*/ to find the end of the code.
-            /__css__\([`'"]((.|\r\n|\r|\n)*?)[`'"](\r\n|\r|\n)\/\*__css_end__\*\/((.|\r\n|\r|\n)*?)\)/g,
-            (match, code) => {
-                const hash = `${utils.getFileHash(filepath)}_${++index}`;
-                const filename = `${hash}.css`;
+            // We're using comment blocks to find the end of the code to extract.
+            /\/\*__reshadow_css_start__\*\/([\s\S]*?)\/\*__reshadow_css_end__\*\//g,
+            (match, codeBlock) => {
+                let [, code] = codeBlock.match(
+                    /__inner_css_start__\*\/([\s\S]*?)\/\*__inner_css_end__/,
+                );
+                code = code.trim().replace(/^[`'"]([\s\S]*?)[`'"]$/, '$1');
 
-                virtualModules.writeModule(
-                    path.resolve(cacheDirectory, filename),
-                    code
-                        .replace(/\\"/g, '"')
-                        .replace(/\\'/g, "'")
-                        .replace(/\\n/g, '\n'),
+                const filepath = options.getFilepath(resourcePath);
+
+                queue.push(
+                    writeModule(
+                        filepath,
+                        code
+                            .replace(/\\"/g, '"')
+                            .replace(/\\'/g, "'")
+                            .replace(/\\n/g, '\n'),
+                    ),
                 );
 
-                return `require('.cache/reshadow/${filename}')`;
+                const [requirePath] = filepath.split('node_modules/').slice(-1);
+
+                return `require('${requirePath}')`;
             },
         )
         .replace(/\/\*__reshadow-styles__:"(.*?)"\*\//, (match, dep) => {
             const depPath = utils.resolveDependency({
                 filename: dep,
-                basedir: path.dirname(filepath),
+                basedir: path.dirname(resourcePath),
             });
 
             this.dependency(depPath);
@@ -73,5 +135,9 @@ module.exports = function(source) {
             return '';
         });
 
-    return result;
-};
+    await Promise.all(queue);
+
+    return [result, inputSourceMap];
+}
+
+module.exports = makeLoader(loader);